Added warning and dangr zone to the SOC vs. Time graph. Added tables to the app for visualization to Mehrdad.
